use url of the context
On request of @thet this is a copy of the original PR from @
spereverde in the *tomgross* repository which is deleted meanwhile. I'm not conviced this patch works for all situations. What if you want to batch a view which is not the standard view. It is very easy to customize the batching view for custom purposes.